requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pypdf>=3.0.0
python-docx>=0.8.11
requests>=2.31.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
